Now checking to see if mgweb process start process errors

+++ Kubernetes/mgwebentry/start.sh
@@ -30,15 +30,17 @@ if [ ! -f "/opt/mgweb/mapped/mpm_event.conf" ]; then
 fi
 
 export ydb_gbldir=/opt/yottadb/yottadb.gld
-ydb <<< "D start^%zmgwebUtils"
-pid=$(lsof | grep mgweb.dat | awk '{ print $2 }')
-if [[ "$pid" != "" ]]
+/opt/mgweb/ydb_run "start^%zmgwebUtils"
+if [[ "$?" != "0" ]]
 then
-	kill -9 $pid
+	pid=$(lsof | grep mgweb.dat | awk '{ print $2 }')
+	if [[ "$pid" != "" ]]
+	then
+		kill -9 $pid
+                /usr/local/lib/yottadb/r130/mupip rundown -region default
+                /opt/mgweb/ydb_run "start^%zmgwebUtils"
+	fi
 fi
-/usr/local/lib/yottadb/r130/mupip rundown -region default
-ydb <<< "D start^%zmgwebUtils"
-/usr/local/lib/yottadb/r130/mupip rundown -region default
 cd /usr/local/YottaDB-dashboard/glbview
 /usr/local/YottaDB-dashboard/glbview/globview.sh start
